% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncrossing_probability.R
\name{compute_noncrossing_prob}
\alias{compute_noncrossing_prob}
\title{Compute the joint probability of ordered random vaiables}
\usage{
compute_noncrossing_prob(h, l, CDF = punif, FFT = TRUE)
}
\arguments{
\item{h}{numeric(n), the higher bound of the ordered random vaiables}

\item{l}{numeric(n), the lower bound of ordered random vaiables}

\item{CDF}{function, a CDF function that specify the distribution of
the random variable. The density of the random variable at x must be
able to be computed via CDF(x).}
}
\description{
Compute the probability P(l_i<X_i<h_i for all i=1,...,N) 
where X_1,..., X_n are ordered continuous random vaiables sampled from a 
distribution
}
