\name{Population_Data}
\alias{Population_Data}
\docType{data}
\title{
The proportion of inhabitants living within a 200 kilometer wide costal strip in 232 countries in the year 2010
}
\description{
This data set contains the proportion of inhabitants living within a 200 kilometer wide costal strip in 232 countries in the year 2010.
In Example 3.3 of Dimitrova, Kaishev, Tan (2017), the data set is modelled using a zero-and-one-inflated beta distribution in the null hypothesis and a one-sample two-sided Kolmogorov-Smirnov test is performed to test whether the proposed distribution fits the data well enough.
}
\usage{data("Population_Data")}
\format{
  A data frame with 232 observations on the proportion of inhabitants living within a 200 kilometer wide costal strip in 2010.
}
\source{
\url{http://sedac.ciesin.columbia.edu/data/set/nagdc-population-landscape-climate-estimates-v3}
}
\references{
Dimitrova D.S, Kaishev V.K, Tan S. (2017) "Computing the Kolmogorov-Smirnov Distribution When the Underlying cdf is Purely Discrete, Mixed or Continuous".
\url{http://openaccess.city.ac.uk/18541}
}
\keyword{datasets}
